+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 FEATURES:
 * storage: add `object_lock_configuration` attribute in `bucket` resource
 * storage: add `object_lock_legal_hold_status`, `object_lock_mode` and `object_lock_retain_until_date` attributes in `object` resource
+* mongodb: add `restore` entity in `yandex_mdb_mongodb_cluster` resource
 
 ## 0.84.0 (December 16, 2022)
 BUG FIXES: